refactor(traits): rename to Has*References with deprecated aliases
